requirements.txt
antlr4_python3_runtime==4.12.0
mypy==1.8.0
